Scripting for seeding

+++ run/finetuning/create.py
@@ -11,6 +11,7 @@ VOCAB_SIZES = [256]
 TOP_KS = [10]
 MODES = ["uni", "bi"]
 TASKS = ["clm", "mlm"]
+SEEDS = [42, 0, 1, 2, 3]
 
 
 JOB_NAME = "JOB_NAME"
@@ -22,15 +23,16 @@ PER_DEVICE_EVAL_BATCH_SIZE = 64
 REPRESENTATION = 8
 ALGORITHM = "ALGORITHM"
 VOCAB_SIZE = "VOCAB_SIZE"
-MAX_LENGTH = 2 ** 12  # FIXME
-DATA_READ_BYTES = 2 ** 12  # FIXME
+SEED = "SEED"
+MAX_LENGTH = 2 ** 16
+DATA_READ_BYTES = 2 ** 16
 TOP_K = "TOP_K"
-BF_OR_FP = "fp"  # FIXME
-TF32 = "false"  # FIXME
+BF_OR_FP = "bf"
+TF32 = "true"
 
-LM_TR_SIZE = 10000 # FIXME
-LM_VL_SIZE = 1000 # FIXME
-LM_SAVE_EVAL_STEPS = 2 # FIXME
+LM_TR_SIZE = 1000000
+LM_VL_SIZE = 10000
+LM_SAVE_EVAL_STEPS = 200
 LM_GRADIENT_ACCUMULATION_STEPS = 8
 
 CLF_GRADIENT_ACCUMULATION_STEPS = 1
@@ -61,6 +63,7 @@ src/learn/train.py \\
 --arch_config='{ARCH_CONFIG}' \\
 --metric_for_best_model="eval_loss" \\
 --task="TASK" \\
+--seed=42 \\
 --streaming=true \\
 --skip_eval_check=false \\
 --dataset_backend="HF" \\
@@ -109,7 +112,7 @@ BODY_CLF = f"""#!/bin/bash -l
 #SBATCH --account=admalware
 #SBATCH --partition=tier3
 #SBATCH --output=./logs/%x_%j.out
-#SBATCH --time=00-12:00:00
+#SBATCH --time=00-6:00:00
 #SBATCH --nodes=1
 #SBATCH --cpus-per-task=1
 #SBATCH --ntasks={4}
@@ -129,6 +132,7 @@ src/learn/train.py \\
 --arch_config='{ARCH_CONFIG}' \\
 --metric_for_best_model="eval_accuracy" \\
 --task="clf" \\
+--seed={SEED} \\
 --streaming=false \\
 --skip_eval_check=false \\
 --top_k={TOP_K} \\
@@ -143,7 +147,7 @@ src/learn/train.py \\
 --output_dir=tmp \\
 --save_strategy="epoch" \\
 --evaluation_strategy="epoch" \\
---num_train_epochs=10 \\
+--num_train_epochs=5 \\
 --logging_steps=10 \\
 --save_steps=100 \\
 --eval_steps=100 \\
@@ -174,8 +178,8 @@ src/learn/train.py \\
 
 OUTPUT = Path(os.path.realpath(__file__)).parent
 PRETRAINED_MODEL_PATH = {  # FIXME
-    "clm": Path("output/finetuning/mamba/representation--8/algorithm--Raw/vocab_size--256/max_length--4096/task--clm/tr_size--10000/depth--1/mode--uni/num_hidden_layers--8/hidden_size--256/embedding_size--256/per_device_train_batch_size--64/gradient_accumulation_steps--8/learning_rate--0.001/weight_decay--0.01/adam_beta1--0.9/adam_beta2--0.999/adam_epsilon--1e-08/max_grad_norm--1.0/lr_scheduler_type--linear/warmup_ratio--0.0/bf16--False/fp16--True/tf32--False/optim--adamw_torch/checkpoints/"),
-    "mlm": Path("output/finetuning/mamba/representation--8/algorithm--Raw/vocab_size--256/max_length--4096/task--mlm/tr_size--10000/depth--1/mode--bi/num_hidden_layers--8/hidden_size--256/embedding_size--256/per_device_train_batch_size--32/gradient_accumulation_steps--16/learning_rate--0.001/weight_decay--0.01/adam_beta1--0.9/adam_beta2--0.999/adam_epsilon--1e-08/max_grad_norm--1.0/lr_scheduler_type--linear/warmup_ratio--0.0/bf16--False/fp16--True/tf32--False/optim--adamw_torch/checkpoints/"),
+    "clm": Path("/home/lk3591/Documents/code/RawByteClf/output/finetuning/mamba/seed-42/representation--8/algorithm--Raw/vocab_size--256/max_length--65536/task--clm/tr_size--1000000/depth--1/mode--uni/num_hidden_layers--8/hidden_size--256/embedding_size--256/per_device_train_batch_size--8/gradient_accumulation_steps--64/learning_rate--0.001/weight_decay--0.01/adam_beta1--0.9/adam_beta2--0.999/adam_epsilon--1e-08/max_grad_norm--1.0/lr_scheduler_type--linear/warmup_ratio--0.0/bf16--True/fp16--False/tf32--True/optim--adamw_torch/checkpoints/"),
+    "mlm": Path("/home/lk3591/Documents/code/RawByteClf/output/finetuning/mamba/seed-42/representation--8/algorithm--Raw/vocab_size--256/max_length--65536/task--mlm/tr_size--1000000/depth--1/mode--bi/num_hidden_layers--8/hidden_size--256/embedding_size--256/per_device_train_batch_size--4/gradient_accumulation_steps--128/learning_rate--0.001/weight_decay--0.01/adam_beta1--0.9/adam_beta2--0.999/adam_epsilon--1e-08/max_grad_norm--1.0/lr_scheduler_type--linear/warmup_ratio--0.0/bf16--True/fp16--False/tf32--True/optim--adamw_torch/checkpoints/"),
 }
 
 
@@ -198,21 +202,30 @@ for algorithm in ALGORITHMS:
                 fp.write(body)
 
             for top_k in TOP_KS:
-                job_name = f"ft-{'clf'}-{mode}-{algorithm[0]}-{vocab_size}-{top_k}"
-                body = BODY_CLF \
-                    .replace("JOB_NAME", job_name) \
-                    .replace("ALGORITHM", algorithm) \
-                    .replace("VOCAB_SIZE", str(vocab_size)) \
-                    .replace("TOP_K", str(top_k))
-                with open(OUTPUT / f"{job_name}.sh", "w") as fp:
-                    fp.write(body)
-
-                checkpoints = list(PRETRAINED_MODEL_PATH[task].iterdir())
-                checkpoints = sorted(checkpoints, key=lambda x: int(x.name.split("-")[1]))
-                for f in checkpoints:
-                    job_name = f"ft-{'ft'}-{f.name.replace('checkpoint-', '')}-{mode}-{algorithm[0]}-{vocab_size}-{top_k}"
-                    body = body \
-                        .replace("mamba", f.as_posix()) \
-                        .replace('"mode": "uni"', f'"mode": "{mode}"')
+                for seed in SEEDS:
+                    job_name = f"ft-{'clf'}-{mode}-{algorithm[0]}-{vocab_size}-{top_k}-{seed}"
+                    body = BODY_CLF \
+                        .replace("JOB_NAME", job_name) \
+                        .replace("ALGORITHM", algorithm) \
+                        .replace("VOCAB_SIZE", str(vocab_size)) \
+                        .replace("TOP_K", str(top_k)) \
+                        .replace("SEED", str(seed))
                     with open(OUTPUT / f"{job_name}.sh", "w") as fp:
                         fp.write(body)
+
+                    if not PRETRAINED_MODEL_PATH[task].exists():
+                        continue
+                    checkpoints = list(PRETRAINED_MODEL_PATH[task].iterdir())
+                    checkpoints = sorted(checkpoints, key=lambda x: int(x.name.split("-")[1]))
+                    for f in checkpoints:
+                        job_name = f"ft-{'ft'}-{f.name.replace('checkpoint-', '')}-{mode}-{algorithm[0]}-{vocab_size}-{top_k}-{seed}"
+                        body = BODY_CLF \
+                            .replace("JOB_NAME", job_name) \
+                            .replace("ALGORITHM", algorithm) \
+                            .replace("VOCAB_SIZE", str(vocab_size)) \
+                            .replace("TOP_K", str(top_k)) \
+                            .replace("SEED", str(seed)) \
+                            .replace("mamba", f'"{f.as_posix()}"') \
+                            .replace('"mode": "uni"', f'"mode": "{mode}"')
+                        with open(OUTPUT / f"{job_name}.sh", "w") as fp:
+                            fp.write(body)
